% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{est_q}
\alias{est_q}
\title{Encontra o quantil da distribuição Qui-Quadrado inf.}
\usage{
est_q(fn, alpha = 0.05, bilateral = FALSE, step = 1e-04, c, k)
}
\arguments{
\item{fn}{Recebe a distribuição Qui-Quadrado inf.}

\item{alpha}{Nível de significância adotado.}

\item{bilateral}{Se \code{TRUE}, retorna os quantis para um teste bilateral. O padrão considera
\code{bilateral  = FALSE}.}

\item{step}{Tamanho do passo da integral numeérica responsável pela obtenção dos quantis da Qui-Quadrado inf.
O padrão considera \code{step = 1e-4}.}

\item{c}{Parâmetro da distribuição Qui-Quadrado inf.}

\item{k}{Parâmetro da distribuição Qui-Quadrado inf.}
}
\description{
Encontra o quantil da distribuição Qui-Quadrado inf.
}
\details{
A função recebe como argumento uma função densidade de probabilidade ou uma função de probabilidade
que é passada como argumento à \code{f}. O objeto passado como argumento de \code{f} deverá ser implementado conforme os
exemplos. Note que se a função for implementada segundo os exemplos, não haverá a necessidade de implementar a função sob a
hipótese nula, visto que são as mesmas funções. Para especificar os parâmetros que serão fixados, i.e, para especificar
a distribuição sob a hipótese nula, utiliza-se o argumento \code{par0} que receberá uma lista formada por dois vetores.
O primeiro vetor da lista de verá ser um vetor de strings com os nomes das variáveis que deseja-se fixar e o segundo vetor
deverá conter os valores que serão atribuidos à cada uma das variáveis passada ao primeiro vetor.
}
\examples{
fdp_chisq_inf <- function(par, x) {
  k <- par[1]
  c <- par[2]
  dchisq(x = x, df = k) * (1 - (1 - pchisq(q = x, df = k)) ^ c + c * pchisq(q = x, df = k) *
                        (1 - pchisq(q = x, df = k)) ^ (c - 1))
}
est_q(fn = fdp_chisq_inf, alpha = 0.05, bilateral = FALSE, c = 1, k = 1)
}
\author{
Pedro Rafael D. Marinho
}
